+++ src/main/java/com/programmers/smrtstore/domain/user/application/vo/DefaultSeparateResult.java
@@ -1,4 +1,4 @@
-package com.programmers.smrtstore.domain.user.application;
+package com.programmers.smrtstore.domain.user.application.vo;
 
 import com.programmers.smrtstore.domain.user.domain.entity.ShippingAddress;
 import java.util.List;
@@ -13,7 +13,8 @@ public class DefaultSeparateResult {
     private ShippingAddress defaultShippingAddress;
     private List<ShippingAddress> notDefaultShippingAddresses;
 
-    public static DefaultSeparateResult of(ShippingAddress shippingAddress, List<ShippingAddress> shippingAddresses) {
+    public static DefaultSeparateResult of(ShippingAddress shippingAddress,
+        List<ShippingAddress> shippingAddresses) {
         return new DefaultSeparateResult(shippingAddress, shippingAddresses);
     }
 }
